Fix params picker when zero params exist
