+++ src/mainviews/create-job.tsx
@@ -3,7 +3,7 @@ import React, { ChangeEvent } from 'react';
 import { Heading } from '../components/heading';
 import { Cluster } from '../components/cluster';
 import { OutputFormatPicker, outputFormatsForEnvironment } from '../components/output-format-picker';
-import { parameterNameMatch, ParametersPicker, parameterValueMatch } from '../components/parameters-picker';
+import { ParametersPicker } from '../components/parameters-picker';
 import { Scheduler, SchedulerService } from '../handler';
 import { useTranslator } from '../hooks';
 import { ICreateJobModel, IOutputFormat } from '../model';
@@ -21,6 +21,26 @@ export interface ICreateJobProps {
   toggleView: () => unknown;
   // Extension point: optional additional component
   advancedOptions: React.ElementType;
+}
+
+function parameterNameMatch(elementName: string): number | null {
+  const parameterNameMatch = elementName.match(/^parameter-(\d+)-name$/);
+
+  if (parameterNameMatch === null) {
+    return null;
+  }
+
+  return parseInt(parameterNameMatch[1]);
+}
+
+function parameterValueMatch(elementName: string): number | null {
+  const parameterValueMatch = elementName.match(/^parameter-(\d+)-value$/);
+
+  if (parameterValueMatch === null) {
+    return null;
+  }
+
+  return parseInt(parameterValueMatch[1]);
 }
 
 export function CreateJob(props: ICreateJobProps): JSX.Element {
